+++ Settings.pm
@@ -23,7 +23,7 @@ sub page {
 sub prefs {
 	return ( $prefs, qw(
 		soloistBin ffmpegBin pythonBin pipewireSink wsPortBase relayPortBase
-		format bitrate deviceNameSuffix apiKey relayBind autostart
+		format bitrate initialVolume deviceNameSuffix apiKey relayBind autostart
 		idleDisconnectSeconds
 	) );
 }